i12e: '/etc/i12e/reboot-required' deleted before triggering reboot (reboot-loop prevention)

+++ internal/pull/pull.go
@@ -19,6 +19,14 @@ function reboot_if_required {
   set -x
   systemctl daemon-reload
   [ -f /etc/systemd/system/k3s.service ] || touch /etc/i12e/k3s-install-required
+  rm -f "$REBOOT_FILE"
+  { set +x; } 2> /dev/null
+  if [ -f "$REBOOT_FILE" ]
+  then
+    echo "error: reboot aborted: cannot delete '$REBOOT_FILE' before 'systemctl reboot' execution"
+    return 1
+  fi
+  set -x
   systemctl reboot
 }
 function pull_if_needed {
@@ -29,7 +37,7 @@ function pull_if_needed {
     echo "pulling artifact.tar.gz provided that '${FLAG_FILE}' doesn't exist..."
     set -x
     rclone cat rem:artifact.tar.gz | tar -C / -xvz
-    touch "${REBOOT_FILE}"
+    touch "$REBOOT_FILE"
     { set +x; } 2>/dev/null
   fi
 }
